+++ R/DataDescriptor.R
@@ -246,12 +246,26 @@ assert_compatible_shapes = function(shapes, dataset) {
   }
 
   iwalk(shapes, function(dataset_shape, name) {
-    if (!is.null(dataset_shape) && !test_equal(shapes[[name]][-1], example[[name]]$shape[-1L])) {
-      expected_shape = example[[name]]$shape
-      expected_shape[1] = NA
-      stopf(paste0("First batch from dataset is incompatible with the provided shape of %s:\n",
-        "* Provided shape: %s.\n* Expected shape: %s."), name,
-        shape_to_str(unname(shapes[name])), shape_to_str(list(expected_shape)))
+    if (is.null(dataset_shape)) {
+      return(NULL)
+    }
+    shape_specified = shapes[[name]]
+    shape_example = example[[name]]$shape
+    if (length(shape_specified) != length(shape_example)) {
+      stopf("The specified number of dimensions for element '%s' is %s, but the dataset returned %s",
+        name, length(shape_specified), length(shape_example))
+    }
+
+    if (all(is.na(shape_specified))) {
+      # compatible with any shape
+      return(NULL)
+    }
+
+    shape_example[is.na(shape_specified)] = NA
+    if (!test_equal(shape_specified, shape_example)) {
+      stopf(paste0("First example batch from dataset is incompatible with the provided shape of %s:\n",
+        "* Observed shape: %s.\n* Specified shape: %s."), name,
+        shape_to_str(example[[name]]$shape), shape_to_str(shape_specified))
     }
   })
 }