+++ R/utils.R
@@ -148,7 +148,7 @@ uniqueify = function(new, existing) {
 
 shape_to_str = function(x) {
   assert(test_list(x) || test_integerish(x) || is.null(x))
-  if (is.numeric(x)) { # single shape
+  if (test_integerish(x)) { # single shape
     return(sprintf("(%s)", paste0(x, collapse = ",")))
   }
   if (is.null(x)) {
@@ -271,37 +271,6 @@ order_named_args = function(f, l) {
   l2
 }
 
-infer_shapes = function(shapes_in, param_vals, output_names, fn, rowwise, id) {
-  sin = shapes_in[[1L]]
-  batch_dim = sin[1L]
-  batchdim_is_unknown = is.na(batch_dim)
-  if (batchdim_is_unknown) {
-    sin[1] = 1L
-  }
-  if (rowwise) {
-    sin = sin[-1L]
-  }
-  tensor_in = mlr3misc::invoke(torch_empty, .args = sin, device = torch_device("cpu"))
-
-  fn_args = names(formals(fn))
-  filtered_params = param_vals[intersect(names(param_vals), fn_args)]
-
-  tensor_out = tryCatch(invoke(fn, tensor_in, .args = filtered_params),
-    error = function(e) {
-      stopf("Input shape '%s' is invalid for PipeOp with id '%s'.", shape_to_str(list(sin)), id)
-    }
-  )
-
-  sout = dim(tensor_out)
-
-  if (rowwise) {
-    sout = c(batch_dim, sout)
-  } else if (batchdim_is_unknown) {
-    sout[1] = NA
-  }
-
-  set_names(list(sout), output_names)
-}
 
 #' @title Network Output Dimension
 #' @description