+++ NEWS.md
@@ -18,6 +18,7 @@
   for generating the `torch::dataset`.
 * fix: `NA` is now a valid shape for lazy tensors
 * feat: lazy tensors of length 0 can now be materialized.
+* feat: Shapes can now have multiple `NA`s and not only the batch dimension can be missing. However, most `nn()` operators still expect only one missing values and will throw an error if multiple dimensions are unknown.
 
 # mlr3torch 0.2.1